integrate mooncake kv store
