+++ mminf/engine/enc_dec_engine.py
@@ -37,6 +37,7 @@ class EncoderDecoderEngine(BaseEngine):
         submodules: dict[str, torch.nn.Module],
         model_config: dict,
         device: torch.device,
+        **kwargs
     ) -> None:
         self.submodules = submodules
         self.device = device
