+++ mminf/engine/flow_engine.py
@@ -32,6 +32,7 @@ class FlowEngine(BaseEngine):
         submodules: dict[str, torch.nn.Module],
         model_config: dict,
         device: torch.device,
+        **kwargs
     ) -> None:
         self.submodules = submodules
         self.device = device
