--- mminf/engine/paged_attention.py
@@ -1,52 +1,0 @@
-from dataclasses import dataclass, field
-import queue
-
-
-@dataclass
-class KVCacheConfig:
-    num_layers: int
-    num_kv_heads: int
-    head_dim: int
-    max_seq_len: int
-    max_num_pages: int = 2048
-    page_size: int = 128
-    num_qo_heads: int | None = None  # Optional, defaults to num_kv_heads
-
-    def __post_init__(self):
-        if self.num_qo_heads is None:
-            self.num_qo_heads = self.num_kv_heads
-
-
-@dataclass
-class KVRequestState:
-    """Per-request KV cache state for the AR engine."""
-    page_indices: list[int] = field(default_factory=list)
-    seq_len: int = 0
-    position_id_start: int = 0
-    is_paused: bool = False
-
-
-class PageAllocator:
-    """Simple page allocator using a FIFO queue of free page indices."""
-
-    def __init__(self, max_num_pages: int):
-        self.max_num_pages = max_num_pages
-        self.free_pages: queue.Queue[int] = queue.Queue()
-        for i in range(max_num_pages):
-            self.free_pages.put(i)
-
-    def allocate(self, n: int) -> list[int]:
-        if self.free_pages.qsize() < n:
-            raise RuntimeError(
-                f"Not enough free pages: requested {n}, "
-                f"available {self.free_pages.qsize()}"
-            )
-        return [self.free_pages.get() for _ in range(n)]
-
-    def free(self, pages: list[int]) -> None:
-        for page in pages:
-            self.free_pages.put(page)
-
-    @property
-    def num_free(self) -> int:
-        return self.free_pages.qsize()
